Implemented the array into checkerboard.cpp and .h
Added the initialization for each method. Next step is to add them into
legal moves, create print list method to be used to print either black
or red's legal moves list. Once these are completed, we will implement
the canCapture method in the else statements that checks if a piece can
be captured.