Applied arrays to regular black pieces
Fixed calculation error in printArray. Mistake on case when we are going
to lower diagonals. Needed to be +1 instead of -1. Arrays work for
regular pieces and printArray is working overall. Next step is to
implement arrays into kinged pieces and after test take away print
statements to clean code up