Case checking and capture cheatlist
I fixed many print erros in getLegalMoves in Checkers.cpp

Implemented:
isBlackEmpty: checks if the legal moves for black is empty. Can be
removed. Was used for testing purposes
isRedEmpty: Same reason as above
isCurrentEmpty: This was to take in the case that someone enters a
current location that is empty. In the getLegalMoves method, it would
cause a segmentation fault so applying this method in the main method to
check prevents this and also allows us to only print out an error
message and ask the user to try again.

Included array to store capture methods. These will be utilized in the
move method to check if the propsed location is in one of our lists. If
it is, we'll move it.

Implemented a cheat list with the possible captures that are available.
Only implemented in the regular red pieces currently but will be added
to the rest of the types in next push

NEXT:
Implement capture method into move method.
Apply cheat list for legal captures to other types
Clean code up to make it easier to read